src/test/java/br/com/cinq/spring/data/sample/application/repository/test/CityRepositoryTest.java
package br.com.cinq.spring.data.sample.application.repository.test;

import static org.assertj.core.api.Assertions.assertThat;
import static org.junit.Assert.assertNotNull;

import java.util.List;

import org.junit.Test;
import org.junit.runner.RunWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase.Replace;
import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
import org.springframework.test.context.ContextConfiguration;
import org.springframework.test.context.junit4.SpringRunner;

import br.com.cinq.spring.data.sample.entity.City;
import br.com.cinq.spring.data.sample.repository.CityRepository;
//import br.com.cinq.spring.data.sample.entity.City;
//import br.com.cinq.spring.data.sample.entity.Country;
//import br.com.cinq.spring.data.sample.repository.CityRepository;

/**
 * Eye candy: implements a sample in using JpaRespositories
 */
@RunWith(SpringRunner.class)
@DataJpaTest
@AutoConfigureTestDatabase(replace=Replace.NONE)
public class CityRepositoryTest {

    @Autowired
    private CityRepository dao;

    @Test
    public void testQueryCity() {

        assertNotNull(dao);
        City city =dao.findById(1L);
        assertThat(city.getName()).isEqualTo("Curitiba");
        List<City> cities = dao.findCitiesByCountryName("France");
        assertThat(cities.size()).isEqualTo(2);
    }
}
